kvm: Move utility/ to types/
FamStruct has been moved to types/ as while it is used quite a bit in
the kvm code, it is certainly not something created for it.

In addition, FileDescriptor has been added to types/ to facilitate both
kvm-related ioctl classes as well as future handle wrappers. In
particular, EventFd, which is used in kvm code, will come from
FileDescriptor.

Also added in types/ is exceptions.hpp, which contains macros for
enabling/disabling exceptions. The appropriate changes in the rest of
the codebase has been made accordingly.