types: Rework FAM structs
FAM structs are now template classes, shifting the runtime heap
allocations we were doing to represent the structs to compile-time stack
allocations (support for heap allocations is planned to be added in the
future). Another benefit of the template design is that many methods in
derived classes could be moved to FamStruct, resulting in simpler
wrappers.

The API has now changed as well due to the new template design. Users
will have to know how much space to allocate for a FAM struct. I don't
really see a situation where you wouldn't know beforehand how many
entries you want, so this isn't much of a problem in practice. As for
knowing how many entries to use for the return type of functions like
`kvm::system::msr_index_list()`, the compiler will tell you or you can
use an `auto` declaration.

The implementation of FamStruct has also changed. Instead of
representing the structs as a buffer of integers, we now leverage the
fact that having 2 members: a FAM struct and a std::array gives us the
desired layout. In addition, this takes care of alignment issues which
weren't considered in the former implementation.