kvm: Revert to non-virtual FamStruct
Turns out, calling pure virtual functions don't resolve to the derived
implementations. So, all methods that depended on member names have been
moved back to the derived classes.

A copy constructor was implemented for Msrs and it seems to be working,
will have to see whether or not it also uses the same allocation methods
as the rhs or not. Also on that note, copy/move constructors have been
deleted in FamStruct, leaving whether they're implemented or not to the
subclasses.