kvm: Add allocator overloads for entry and iterator FamStruct constructors
The calls to m_alloc.construct() and m_alloc.destroy() are removed since
the structs wrapped by FamStruct are POD and therefore don't have any
constructors to call; a memset is sufficient for construction.

Non-default constructors of subclasses of FamStruct now call the
appropriate overloads to initialize their base class.

Change Base::begin() to begin() since that method moved to the
subclasses.