kvm: Remove system::get_list()
Default construction no longer sets nmsrs/nent/nr to N, but 0. And
because the ioctls which get_list() is used for check whether nmsrs is
large enough to contain the number of to-be-returned entries, we need
some way of setting that field appropriately, which we could not do in
get_list().