kvm: Add deduction for FamStruct parameters
By using `auto` to do non-type template parameter deduction, we can have
subclasses of FamStruct omit the type of the struct they're wrapping.
It simplifies deduction for the data member pointers as well.

Furthermore, subclasses may omit the type of elements in the FAM thanks
to some tricks with `declval`.